Revert "Benchmark functions were locked into one of the tests"
